+++ module/Core/src/Visit/VisitsStatsHelperInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Core\Visit;
+
+use Shlinkio\Shlink\Core\Visit\Model\VisitsStats;
+
+interface VisitsStatsHelperInterface
+{
+    public function getVisitsStats(): VisitsStats;
+}