+++ module/Rest/src/Action/ShortUrl/ResolveShortUrlAction.php
@@ -7,7 +7,6 @@ namespace Shlinkio\Shlink\Rest\Action\ShortUrl;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
-use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Core\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\Core\Service\ShortUrl\ShortUrlResolverInterface;
 use Shlinkio\Shlink\Core\Transformer\ShortUrlDataTransformer;
@@ -21,12 +20,8 @@ class ResolveShortUrlAction extends AbstractRestAction
     private ShortUrlResolverInterface $urlResolver;
     private array $domainConfig;
 
-    public function __construct(
-        ShortUrlResolverInterface $urlResolver,
-        array $domainConfig,
-        ?LoggerInterface $logger = null
-    ) {
-        parent::__construct($logger);
+    public function __construct(ShortUrlResolverInterface $urlResolver, array $domainConfig)
+    {
         $this->urlResolver = $urlResolver;
         $this->domainConfig = $domainConfig;
     }